tools/onboard/requirements.txt
cantools>=39.0
jinja2>=3.1
pyyaml>=6.0
jsonschema>=4.20
